Merge with rootston.
Enable C++ for bits of the code, but nothing that includes wlroots as that
doesn't want to build as C++.

Enable SubstructureRedirectMask mode and handle MapRequest for now; this lets
us get the WL_SURFACE_ID ClientMessages.

This uses a build of wlroots that has its internal XWM ripped out.