steamcompmgr: get rid of set_win_hidden
The HIDDEN state was originally just a hint for Panorama windows. It can confuse
other clients. The other states can actually be harmful too, like setting
FULLSCREEN was technically incorrect in many cases.

We might want to always set the FOCUS state on every window, but we never needed
it before so I'm inclined to wait and see there.