steamcompmgr: Track X dirtying outside of dispatch
If we dirty the X server outside of the dispatch, xlib can flush its events, leading us to a state where XPending(dpy), would be true but the poll returns false.

Check if we have locally queued events before polling and flush those to mitigate this. This does not involve a system call.