drm: Disable BLEND_TF + REGAMMA_TF if single plane
BLEND_TF with de-pq currently introduces very slight banding on bright highlights of PQ content (eg. sun in Ori and the Will of the Wisps)

As a workaround, let's disable BLEND_TF + REGAMMA_TF if we are doing single plane.

Disabling un-necessary blocks is probably a good thing in the long run anyway for multiple reasons.