renderervulkan: only consider modifiers support all the image properties
If you have a modifier with image compression that is only support for
some image formats like :
   - B8G8R8A8_UNORM : compression supported
   - B8G8R8A8_SRB   : compression not supported

Then the render will not use compression but the modifier indicates it
is present, so the compositor will try to make use of the side
compressed data and this will lead to corruptions.

This fixes issues on Intel HW of Gfx9 generations.