meson: Attempt to deter force_fallback_for being unset
Packagers have been overriding our `force_fallback_for` definition, and replacing our submodules with their system libraries.

Gamescope regularly takes untagged work, and sometimes has its own forks/branches of projects with fixes and features that we need.

We also don't religiously follow wlroots versioning, which does not have a stable ABI, which is problematic when people want to mix and match.

Seriously, you should not change this! It is there for a reason, and I am also tired of debugging random issues reported across a myriad of distros with differing random patches because of this.