rendervulkan: Fix a CVulkanCmdBuffer leak that could result in screenshot request failures
CVulkanDevice::resetCmdBuffers expects m_pendingCmdBufs to be sorted.
Using an unordered_map can result in cases where we eagerly exit out of
the loop without fully cleaning all the relevant resources. This will
result in leaking the command buffer and any resources it transitively
references.

This can be a problem for screenshots as there are only two screenshot
image slots available. If we leak references to these images, at some
point we won't be able to allocate new images for screenshot capture and
the operation will fail.